remove encoding type from heaer field in schema

+++ src/main/java/org/coffee/domain/beans/Coffee.java
@@ -4,10 +4,10 @@ public class Coffee {
     private final String name;
     private final String order_path;
     private final double price;
-    private final double caffeine_level;
+    private final int caffeine_level;
     private final double milk_ratio;
 
-    public Coffee(String name, String order_path, double price, double caffeine_level, double milk_level) {
+    public Coffee(String name, String order_path, double price, int caffeine_level, double milk_level) {
         this.name = name;
         this.order_path = order_path;
         this.price = price;
